LEA: Split up the file `device_panel.rs` into multiple separate files for each component.
