+++ opendut-lea/src/peers/configurator/tabs/devices/name_input.rs
@@ -1,0 +1,62 @@
+use leptos::prelude::*;
+use opendut_lea_components::{UserInput, UserInputValue};
+use opendut_model::topology::{DeviceName, IllegalDeviceName};
+use crate::peers::configurator::types::devices::UserDeviceConfiguration;
+use crate::peers::configurator::types::EMPTY_DEVICE_NAME_ERROR_MESSAGE;
+
+#[component]
+pub fn DeviceNameInput(
+    device_configuration: RwSignal<UserDeviceConfiguration>,
+) -> impl IntoView {
+
+    let (getter, setter) = create_slice(device_configuration,
+        |device_configuration| {
+            Clone::clone(&device_configuration.name)
+        },
+        |device_configuration, value| {
+            device_configuration.name = value;
+        }
+    );
+
+    let validator = |input: String| {
+        match DeviceName::try_from(input.clone()) {
+            Ok(_) => {
+                UserInputValue::Right(input)
+            }
+            Err(cause) => {
+                match cause {
+                    IllegalDeviceName::TooShort { expected, actual, value } => {
+                        if actual > 0 {
+                            UserInputValue::Both(format!("A device name must be at least {expected} characters long."), value)
+
+                        }
+                        else {
+                            UserInputValue::Both(String::from(EMPTY_DEVICE_NAME_ERROR_MESSAGE), value)
+                        }
+                    },
+                    IllegalDeviceName::TooLong { expected, value, .. } => {
+                        UserInputValue::Both(format!("A device name must be at most {expected} characters long."), value)
+                    },
+                    IllegalDeviceName::InvalidStartEndCharacter { value } => {
+                        UserInputValue::Both("The device name starts/ends with an invalid character. \
+                        Valid characters are a-z, A-Z and 0-9.".to_string(), value)
+                    }
+                    IllegalDeviceName::InvalidCharacter { value } => {
+                        UserInputValue::Both("The device name contains invalid characters. \
+                        Valid characters are a-z, A-Z, 0-9 and _-".to_string(), value)
+                    }
+                }
+            }
+        }
+    };
+
+    view! {
+        <UserInput
+            getter
+            setter
+            label="Name"
+            placeholder="Device_A"
+            validator
+        />
+    }
+}
