+++ opendut-lea/src/peers/configurator/tabs/devices/tag_input.rs
@@ -1,0 +1,8 @@
+use leptos::prelude::*;
+
+#[component]
+pub fn DeviceTagInput() -> impl IntoView {
+    view! {
+
+    }
+}
